requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pandas>=1.5.0
matplotlib>=3.6.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
streamlit>=1.25.0
Pillow>=8.0.0
lap>=0.5.12
scipy>=1.9.0